Sage: Refactor bankcard, vault, and virtual check into one class
Closes #1978
