+++ CHANGELOG
@@ -21,6 +21,7 @@
 * Element: Update the live URL endpoint [davidsantoso]
 * Element: Parse responses from unexpected API errors [davidsantoso]
 * Transfirst: Remove unused fields for echeck [davidsantoso]
+* Sage: Internal refactoring into a single gateway class w/ common http conn [anellis]
 
 
 == Version 1.56.0 (December 1, 2015)
