+++ test/fixtures.yml
@@ -866,8 +866,8 @@ s5:
   password: n3yNMBGK
 
 sage:
-  login: login
-  password: password
+  login: 214282982451
+  password: 'Z5W2S8J7X8T5'
 
 sage_pay:
   login: LOGIN
