+++ test/unit/gateways/sage_test.rb
@@ -1,21 +1,33 @@
 require 'test_helper'
 
-class SageBankcardGatewayTest < Test::Unit::TestCase
+class SageGatewayTest < Test::Unit::TestCase
   include CommStub
 
   def setup
-    @gateway = SageBankcardGateway.new(
+    @gateway = SageGateway.new(
                  :login => 'login',
                  :password => 'password'
                )
 
     @credit_card = credit_card
+    @check = check
     @amount = 100
 
     @options = {
       :order_id => '1',
       :billing_address => address,
       :description => 'Store Purchase'
+    }
+
+    @check_options = {
+      :order_id => generate_unique_id,
+      :billing_address => address,
+      :shipping_address => address,
+      :email => 'longbob@example.com',
+      :drivers_license_state => 'CA',
+      :drivers_license_number => '12345689',
+      :date_of_birth => Date.new(1978, 8, 11),
+      :ssn => '078051120'
     }
   end
 
@@ -207,6 +219,113 @@ class SageBankcardGatewayTest < Test::Unit::TestCase
     assert_equal "Outside of United States", post[:C_state]
   end
 
+  def test_successful_check_purchase
+    @gateway.expects(:ssl_post).returns(successful_check_purchase_response)
+
+    response = @gateway.purchase(@amount, @check, @check_options)
+    assert_instance_of Response, response
+    assert_success response
+
+    assert_equal "ACCEPTED", response.message
+    assert_equal "C5O8NUdNt0;virtual_check", response.authorization
+
+    assert_equal "A",                    response.params["success"]
+    assert_equal "",                     response.params["code"]
+    assert_equal "ACCEPTED",             response.params["message"]
+    assert_equal "00",                   response.params["risk"]
+    assert_equal "C5O8NUdNt0",           response.params["reference"]
+    assert_equal "89be635e663b05eca587", response.params["order_number"]
+    assert_equal  "0",                   response.params["authentication_indicator"]
+    assert_equal  "NONE",                response.params["authentication_disclosure"]
+  end
+
+  def test_declined_check_purchase
+    @gateway.expects(:ssl_post).returns(declined_check_purchase_response)
+
+    response = @gateway.purchase(@amount, @check, @check_options)
+    assert_failure response
+    assert response.test?
+    assert_equal "INVALID C_RTE", response.message
+    assert_equal "C5O8NR6Nr0;virtual_check", response.authorization
+
+    assert_equal "X",                    response.params["success"]
+    assert_equal "900016",               response.params["code"]
+    assert_equal "INVALID C_RTE",        response.params["message"]
+    assert_equal "00",                   response.params["risk"]
+    assert_equal "C5O8NR6Nr0",           response.params["reference"]
+    assert_equal "d98cf50f7a2430fe04ad", response.params["order_number"]
+    assert_equal  "0",                    response.params["authentication_indicator"]
+    assert_equal  nil,                    response.params["authentication_disclosure"]
+  end
+
+  def test_successful_store
+    response = stub_comms do
+      @gateway.store(@credit_card, @options)
+    end.check_request do |endpoint, data, headers|
+      assert_match(/<ns1:M_ID>login<\/ns1:M_ID>/, data)
+      assert_match(/<ns1:M_KEY>password<\/ns1:M_KEY>/, data)
+      assert_match(/<ns1:CARDNUMBER>#{credit_card.number}<\/ns1:CARDNUMBER>/, data)
+      assert_match(/<ns1:EXPIRATION_DATE>#{expected_expiration_date}<\/ns1:EXPIRATION_DATE>/, data)
+      assert_equal headers['SOAPAction'], 'https://www.sagepayments.net/web_services/wsVault/wsVault/INSERT_CREDIT_CARD_DATA'
+    end.respond_with(successful_store_response)
+
+    assert response
+    assert_instance_of Response, response
+    assert_success response
+    assert_equal '66234d2dfec24efe9fdcd4b751578c11', response.authorization
+    assert_equal 'Success', response.message
+  end
+
+  def test_failed_store
+    response = stub_comms do
+      @gateway.store(@credit_card, @options)
+    end.check_request do |endpoint, data, headers|
+      assert_match(/<ns1:M_ID>login<\/ns1:M_ID>/, data)
+      assert_match(/<ns1:M_KEY>password<\/ns1:M_KEY>/, data)
+      assert_match(/<ns1:CARDNUMBER>#{credit_card.number}<\/ns1:CARDNUMBER>/, data)
+      assert_match(/<ns1:EXPIRATION_DATE>#{expected_expiration_date}<\/ns1:EXPIRATION_DATE>/, data)
+      assert_equal headers['SOAPAction'], 'https://www.sagepayments.net/web_services/wsVault/wsVault/INSERT_CREDIT_CARD_DATA'
+    end.respond_with(failed_store_response)
+
+    assert response
+    assert_instance_of Response, response
+    assert_failure response
+    assert_nil response.authorization
+    assert_equal 'Unable to verify vault service', response.message
+  end
+
+  def test_successful_unstore
+    response = stub_comms do
+      @gateway.unstore('1234', @options)
+    end.check_request do |endpoint, data, headers|
+      assert_match(/<ns1:M_ID>login<\/ns1:M_ID>/, data)
+      assert_match(/<ns1:M_KEY>password<\/ns1:M_KEY>/, data)
+      assert_match(/<ns1:GUID>1234<\/ns1:GUID>/, data)
+      assert_equal headers['SOAPAction'], 'https://www.sagepayments.net/web_services/wsVault/wsVault/DELETE_DATA'
+    end.respond_with(successful_unstore_response)
+
+    assert response
+    assert_instance_of Response, response
+    assert_success response
+    assert_equal 'Succeeded', response.message
+  end
+
+  def test_failed_unstore
+    response = stub_comms do
+      @gateway.unstore('1234', @options)
+    end.check_request do |endpoint, data, headers|
+      assert_match(/<ns1:M_ID>login<\/ns1:M_ID>/, data)
+      assert_match(/<ns1:M_KEY>password<\/ns1:M_KEY>/, data)
+      assert_match(/<ns1:GUID>1234<\/ns1:GUID>/, data)
+      assert_equal headers['SOAPAction'], 'https://www.sagepayments.net/web_services/wsVault/wsVault/DELETE_DATA'
+    end.respond_with(failed_unstore_response)
+
+    assert response
+    assert_instance_of Response, response
+    assert_failure response
+    assert_equal 'Failed', response.message
+  end
+
   private
   def successful_authorization_response
     "\002A911911APPROVED                        00MX001234567890\0341000\0340\034\003"
@@ -234,5 +353,91 @@ class SageBankcardGatewayTest < Test::Unit::TestCase
 
   def invalid_login_response
     "\002X911911SECURITY VIOLATION              00P 000000000000\034\0340\034\003"
+  end
+
+  def successful_check_purchase_response
+    "\002A      ACCEPTED                        00C5O8NUdNt0\03489be635e663b05eca587\0340\034NONE\034\003"
+  end
+
+  def declined_check_purchase_response
+    "\002X900016INVALID C_RTE                   00C5O8NR6Nr0\034d98cf50f7a2430fe04ad\0340\034\034\003"
+  end
+
+  def expected_expiration_date
+    '%02d%02d' % [@credit_card.month, @credit_card.year.to_s[2..4]]
+  end
+
+  def successful_store_response
+    <<-XML
+<?xml version="1.0" encoding="utf-8" ?>
+<soap:Envelope xmlns:soap="http://www.w3.org/2003/05/soap-envelope" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <soap:Body>
+    <INSERT_CREDIT_CARD_DATAResponse xmlns="https://www.sagepayments.net/web_services/wsVault/wsVault">
+      <INSERT_CREDIT_CARD_DATAResult>
+        <!-- Bunch of xs:schema stuff. Then... -->
+        <diffgr:diffgram xmlns:msdata="urn:schemas-microsoft-com:xml-msdata" xmlns:diffgr="urn:schemas-microsoft-com:xml-diffgram-v1">
+          <NewDataSet xmlns="">
+            <Table1 diffgr:id="Table11" msdata:rowOrder="0" diffgr:hasChanges="inserted">
+              <SUCCESS>true</SUCCESS>
+              <GUID>66234d2dfec24efe9fdcd4b751578c11</GUID>
+              <MESSAGE>SUCCESS</MESSAGE>
+            </Table1>
+          </NewDataSet>
+        </diffgr:diffgram>
+      </INSERT_CREDIT_CARD_DATAResult>
+    </INSERT_CREDIT_CARD_DATAResponse>
+  </soap:Body>
+</soap:Envelope>
+    XML
+  end
+
+  def failed_store_response
+    <<-XML
+<?xml version="1.0" encoding="utf-8" ?>
+<soap:Envelope xmlns:soap="http://www.w3.org/2003/05/soap-envelope" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <soap:Body>
+    <INSERT_CREDIT_CARD_DATAResponse xmlns="https://www.sagepayments.net/web_services/wsVault/wsVault">
+      <INSERT_CREDIT_CARD_DATAResult>
+        <!-- Bunch of xs:schema stuff. Then... -->
+        <diffgr:diffgram xmlns:msdata="urn:schemas-microsoft-com:xml-msdata" xmlns:diffgr="urn:schemas-microsoft-com:xml-diffgram-v1">
+          <NewDataSet xmlns="">
+            <Table1 diffgr:id="Table11" msdata:rowOrder="0" diffgr:hasChanges="inserted">
+              <SUCCESS>false</SUCCESS>
+              <GUID />
+              <MESSAGE>UNABLE TO VERIFY VAULT SERVICE</MESSAGE>
+            </Table1>
+          </NewDataSet>
+        </diffgr:diffgram>
+      </INSERT_CREDIT_CARD_DATAResult>
+    </INSERT_CREDIT_CARD_DATAResponse>
+  </soap:Body>
+</soap:Envelope>
+    XML
+  end
+
+  def successful_unstore_response
+    <<-XML
+<?xml version="1.0" encoding="utf-8" ?>
+<soap:Envelope xmlns:soap="http://www.w3.org/2003/05/soap-envelope" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <soap:Body>
+    <DELETE_DATAResponse xmlns="https://www.sagepayments.net/web_services/wsVault/wsVault">
+      <DELETE_DATAResult>true</DELETE_DATAResult>
+    </DELETE_DATAResponse>
+  </soap:Body>
+</soap:Envelope>
+    XML
+  end
+
+  def failed_unstore_response
+    <<-XML
+<?xml version="1.0" encoding="utf-8" ?>
+<soap:Envelope xmlns:soap="http://www.w3.org/2003/05/soap-envelope" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xmlns:xsd="http://www.w3.org/2001/XMLSchema">
+  <soap:Body>
+    <DELETE_DATAResponse xmlns="https://www.sagepayments.net/web_services/wsVault/wsVault">
+      <DELETE_DATAResult>false</DELETE_DATAResult>
+    </DELETE_DATAResponse>
+  </soap:Body>
+</soap:Envelope>
+    XML
   end
 end
